OBGM-47 Refactor ProductAttributeEcleImporter
